API: use uint64_t for sizes that can exceed 32-bit range
Change size_t to uint64_t in two public API locations that could
overflow on 32-bit platforms with large images:

- fsd_patch_output_size(): size_out parameter (size_t* → uint64_t*)
- fsd_diff_stats_t::patch_size field (size_t → uint64_t)

These represent file/image sizes that are not bounded by the platform's
address space, so uint64_t is the correct type. Fixing before 1.0.0
release to avoid a breaking API change later.

+++ include/fsdiff/patch.h
@@ -61,7 +61,7 @@ fsd_error_t fsd_patch_read_header(const char *patch_path,
  * @return            FSD_SUCCESS or error code
  */
 fsd_error_t fsd_patch_output_size(const char *patch_path,
-                                  size_t *size_out);
+                                  uint64_t *size_out);
 
 /**
  * Set progress callback for patch operations.
